+++ src/main/groovy/rocks/metaldetector/butler/service/importjob/MetalArchivesReleaseImporter.groovy
@@ -1,4 +1,4 @@
-package rocks.metaldetector.butler.service.release
+package rocks.metaldetector.butler.service.importjob
 
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
