+++ src/main/groovy/rocks/metaldetector/butler/web/dto/ImportJobResponse.groovy
@@ -12,5 +12,6 @@ class ImportJobResponse {
   int totalCountImported
   LocalDateTime startTime
   LocalDateTime endTime
+  String source
 
 }
