+++ src/test/groovy/rocks/metaldetector/butler/service/release/PersistReleaseEntityTaskTest.groovy
@@ -3,6 +3,7 @@ package rocks.metaldetector.butler.service.release
 import rocks.metaldetector.butler.model.release.ReleaseEntity
 import rocks.metaldetector.butler.model.release.ReleaseRepository
 import rocks.metaldetector.butler.service.cover.CoverService
+import rocks.metaldetector.butler.service.importjob.PersistReleaseEntityTask
 import spock.lang.Specification
 
 class PersistReleaseEntityTaskTest extends Specification {
