+++ src/test/groovy/rocks/metaldetector/butler/service/release/ReleaseServiceTest.groovy
@@ -2,11 +2,8 @@ package rocks.metaldetector.butler.service.release
 
 import org.springframework.data.domain.PageImpl
 import rocks.metaldetector.butler.model.TimeRange
-import rocks.metaldetector.butler.model.importjob.ImportJobEntity
-import rocks.metaldetector.butler.model.importjob.ImportJobRepository
 import rocks.metaldetector.butler.model.release.ReleaseEntity
 import rocks.metaldetector.butler.model.release.ReleaseRepository
-import rocks.metaldetector.butler.web.dto.CreateImportJobResponse
 import rocks.metaldetector.butler.web.dto.ReleaseDto
 import spock.lang.Specification
 import spock.lang.Unroll
@@ -15,71 +12,14 @@ import java.time.LocalDate
 
 import static rocks.metaldetector.butler.DtoFactory.ReleaseDtoFactory
 import static rocks.metaldetector.butler.DtoFactory.ReleaseEntityFactory
-import static rocks.metaldetector.butler.model.release.ReleaseSource.METAL_ARCHIVES
-import static rocks.metaldetector.butler.model.release.ReleaseSource.METAL_HAMMER_DE
 
 class ReleaseServiceTest extends Specification {
 
   ReleaseServiceImpl underTest = new ReleaseServiceImpl(
-      releaseRepository: Mock(ReleaseRepository),
-      importJobRepository: Mock(ImportJobRepository),
-      metalArchivesReleaseImportService: Mock(MetalArchivesReleaseImporter),
-      metalHammerReleaseImportService: Mock(MetalHammerReleaseImporter)
+      releaseRepository: Mock(ReleaseRepository)
   )
 
   static LocalDate NOW = LocalDate.now()
-
-  def "importFromExternalSources: should create a new import job before start importing new releases"() {
-    when:
-    underTest.importFromExternalSources()
-
-    then:
-    1 * underTest.importJobRepository.save({
-      assert it.jobId != null
-      assert it.startTime != null
-      assert it.source == METAL_ARCHIVES
-    }) >> new ImportJobEntity(jobId: UUID.randomUUID())
-
-    then:
-    1 * underTest.metalArchivesReleaseImportService.importReleases(*_)
-
-    then:
-    1 * underTest.importJobRepository.save({
-      assert it.jobId != null
-      assert it.startTime != null
-      assert it.source == METAL_HAMMER_DE
-    }) >> new ImportJobEntity(jobId: UUID.randomUUID())
-
-    then:
-    1 * underTest.metalHammerReleaseImportService.importReleases(*_)
-  }
-
-  def "importFromExternalSources: should pass internal import job id to metalArchivesReleaseImportService"() {
-    given:
-    Long id = 666
-    underTest.importJobRepository.save(*_) >> new ImportJobEntity(id: id)
-
-    when:
-    underTest.importFromExternalSources()
-
-    then:
-    1 * underTest.metalArchivesReleaseImportService.importReleases(id)
-
-    and:
-    1 * underTest.metalHammerReleaseImportService.importReleases(id)
-  }
-
-  def "importFromExternalSources: should return response with import job id"() {
-    given:
-    UUID jobId = UUID.randomUUID()
-    underTest.importJobRepository.save(*_) >> new ImportJobEntity(jobId: jobId)
-
-    when:
-    def result = underTest.importFromExternalSources()
-
-    then:
-    result == new CreateImportJobResponse(jobIds: [jobId, jobId])
-  }
 
   def "find all upcoming releases for #artists (paginated)"() {
     given:
